pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anteroom"
version = "0.15.1"
description = "Anteroom - your gateway to AI conversation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sse-starlette>=1.8.0",
    "openai>=1.12.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "filetype>=1.2.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "tiktoken>=0.7.0",
    "cryptography>=42.0.0",
    "argon2-cffi>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0",
    "httpx>=0.25.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pip-audit>=2.7.0",
]

[project.scripts]
aroom = "anteroom.__main__:main"
anteroom = "anteroom.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
anteroom = ["static/**/*", "cli/default_skills/*.yaml"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["anteroom"]

[tool.coverage.report]
show_missing = true
